Rename package to bible addict
