+++ app/src/androidTest/java/com/junaidi/BibleAddict/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.junaidi.qtdays
+package com.junaidi.BibleAddict
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.junaidi.qtdays", appContext.packageName)
+        assertEquals("com.junaidi.BibleAddict", appContext.packageName)
     }
 }
