+++ app/src/main/java/com/junaidi/BibleAddict/MainActivity.kt
@@ -1,28 +1,20 @@
-package com.junaidi.qtdays
+package com.junaidi.BibleAddict
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material.Checkbox
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.junaidi.qtdays.ui.theme.BibleAddictTheme
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import com.junaidi.qtdays.Models.Profile
-import com.junaidi.qtdays.maincomposables.Heading
-import com.junaidi.qtdays.maincomposables.SingleCheckbox
-import com.junaidi.qtdays.maincomposables.WeeklyCheckboxes
-import java.util.Calendar
+import com.junaidi.BibleAddict.ui.theme.BibleAddictTheme
+import com.junaidi.BibleAddict.Models.Profile
+import com.junaidi.BibleAddict.maincomposables.Heading
+import com.junaidi.BibleAddict.maincomposables.WeeklyCheckboxes
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
