+++ app/src/main/java/com/junaidi/BibleAddict/Models/Profile.kt
@@ -1,3 +1,3 @@
-package com.junaidi.qtdays.Models
+package com.junaidi.BibleAddict.Models
 
 data class Profile(val currency: Int, val streak: Int, val name: String)
