+++ app/src/main/java/com/junaidi/BibleAddict/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.junaidi.qtdays.ui.theme
+package com.junaidi.BibleAddict.ui.theme
 
 import androidx.compose.ui.graphics.Color
