+++ app/src/test/java/com/junaidi/BibleAddict/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.junaidi.qtdays
+package com.junaidi.BibleAddict
 
 import org.junit.Test
